+++ src/pages/Survey/Survey.css
@@ -1,5 +1,6 @@
 .Survey-container {
   margin: auto;
+  margin-top: 30px;
   height: 100%;
   width: 600px;
   border: 1px solid lightgrey;
