+++ src/pages/Survey/SurveyResult.jsx
@@ -1,69 +1,78 @@
-import useNavi from "../hooks/useNavi";
-
+import useNavi from "../../hooks/useNavi";
+import './SurveyResult.css';
 const SurveyResult = ({topTwoGenres}) =>{
 
    const { goTo } = useNavi();
   const genreCombinations = {
   "Action+Simulation": {
-    title: "총질 마스터 🔫🧠",
+    title: "총질 마스터",
     desc: "빠른 손가락 + 계산적인 머리, FPS 속의 참모장교",
   },
   "Action+Racing": {
-    title: "질주광 🚗💥",
+    title: "질주광",
     desc: "속도는 곧 생명, 브레이크 따위는 장식",
   },
   "Action+Sports": {
-    title: "운동 돌격대 🏀🥊",
+    title: "운동 돌격대",
     desc: "필드 위 전투요정, 스포츠판 람보",
   },
   "Action+RPG": {
-    title: "모험돌이 🗡️🐉",
+    title: "모험돌이",
     desc: "던전도, 드래곤도 그냥 오늘의 점심거리",
   },
   "Simulation+Racing": {
-    title: "가상 레이싱왕 🏎️🕹️",
+    title: "가상 레이싱왕",
     desc: "F1 드라이버를 꿈꾸는 방구석 설계자",
   },
   "Simulation+Sports": {
-    title: "전략왕 스포츠맨 📊⚽",
+    title: "전략왕 스포츠맨",
     desc: "선수도 감독도 내가 한다! 승부조작(?) 금지",
   },
   "Simulation+RPG": {
-    title: "생활왕 RPG 🏡🧙",
+    title: "생활왕 RPG",
     desc: "농사도 하고 몬스터도 때려잡는, 1인 다역 인생왕",
   },
   "Racing+Sports": {
-    title: "트랙의 지배자 🏁⚡",
+    title: "트랙의 지배자",
     desc: "잔디든 아스팔트든 다 내 경기장",
   },
   "Racing+RPG": {
-    title: "질주 모험러 🛞🗺️",
+    title: "질주 모험러",
     desc: "바람보다 빠르게, 퀘스트보다 멀리",
   },
   "Sports+RPG": {
-    title: "챔피언 탐험가 🏆🗡️",
+    title: "챔피언 탐험가",
     desc: "리그 우승 + 세계 구원, 하루 2승 클리어",
   },
 };
+  // 장르 조합하기
   const combinationKey = topTwoGenres
-    ? [...topTwoGenres].sort().join('+')
+    ? Object.keys(genreCombinations).find(key =>
+        topTwoGenres.every(genre => key.includes(genre))
+      )
     : null;
 
   const result = genreCombinations[combinationKey];
 
   return(
-    <div>
+    <div className="items-container">
       {result ? (
-          <div className="items-center">
+          <div className="items">
             <h2>{result.title}</h2>
             <p>{result.desc}</p>
+            <img src={`../src/images/${result.title}.PNG`} alt={result.title} />
           </div>
         ) : (
           <p>결과를 찾을 수 없습니다. 다시 시도해 주세요.</p>
         )}
-        <button onClick={()=>{
+        <div className="btn-container">
+         <button className="btn1" onClick={()=>{
           goTo('/')
         }}>처음으로 돌아가기</button>
+
+          <button className="btn2" onClick={() => {
+          }}>내게 맞는 게임 보기</button>
+        </div>
     </div>
   )
 }
